requirements.txt
Adafruit-Blinka==8.20.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-requests==2.0.0
adafruit-circuitpython-scd4x==1.3.9
adafruit-circuitpython-typing==1.9.4
Adafruit-PlatformDetect==3.47.0
Adafruit-PureIO==1.1.11
certifi==2023.5.7
charset-normalizer==3.2.0
idna==3.4
influxdb==5.3.1
msgpack==1.0.5
pyftdi==0.54.0
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3
pyusb==1.2.1
PyYAML==6.0
requests==2.31.0
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
six==1.16.0
sysv-ipc==1.1.0
typing-extensions==4.7.1
urllib3==2.0.3
